+++ server/component/style/surveyJS/js/4_surveyJS.js
@@ -12,9 +12,6 @@ function initSurveyJS() {
         const surveyContent = $(this).data('survey-js');
         const surveyFields = $(this).data('survey-js-fields');
         const lastResponse = $(this).data('survey-js-last-response');
-        console.log(surveyContent);
-        console.log(surveyFields);
-        console.log(lastResponse);
         $(this).removeAttr('data-survey-js');
         $(this).removeAttr('data-survey-js-fields');
         $(this).removeAttr('data-survey-js-last-response');
@@ -41,7 +38,6 @@ function initSurveyJS() {
                 if (!timeoutExpired) {
                     // load the survey if not expired based on the configuration
                     survey.mergeData(notCompletedSurvey);
-                    console.log(survey.data);
                     survey.setValue('trigger_type', 'updated');
                     if (survey.data.pageNo) {
                         survey.currentPageNo = survey.data.pageNo;