+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - load dropdown values from restful service. When creating a page for the survey it should be advanced example: /survey-js/`[v:data]?`where data will be variable name that will search for the retrieved data. The retrieved data should be in JSON format. Do not use scope in the data config. Then in SurveyJS, go to `Choices from RESTful service` set the url of the survey where is, eg: `test.com/survey-js/Task` where the survey is on page survey-js and `Task` is the name of the table. Then in the `values` set the table columns which should be the value and then in `display text` set the column for display text.
  - always start survey form page `0` on edit mode when is in status `finished`
  - add field `dynamic_replacement` where the survey JSON can be copied and then the mapper can be used for complex dynamic replacements. If the field is set, it will be used, otherwise the Survey js is pulled from the selected survey from the dropdown
+ - add `resetOnBack` property for SurveyJS Pages. When enabled, all answers on this page will be cleared when the user navigates back to it.
 
 
 # v1.3.11