+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v1.4.0 Requires SelfHelp v7.3.1+ Not published
+# v1.4.0 Requires SelfHelp v7.3.1+
 ### New Features
  - load survey in edit mode
  - add `own_entries_only` to style `surveyJS`. If enabled a person can edit only their own responses, otherwise they can edit a foreign response if they have access